fix(ci): resolve all remaining clippy warnings
- Fix approximate constant warnings: Replace 3.14/2.71828 with std::f32/f64::consts::PI/E
- Fix manual C string literals: Replace b"string\0".as_ptr() with c"string".as_ptr()
- Fix bool assertion warnings: Replace assert_eq!(x, true/false) with assert!(...)
- Fix miscellaneous warnings:
  * manual range contains -> use (min..=max).contains(&val)
  * redundant closures -> direct function references
  * len comparison to zero -> use !is_empty()
  * useless vec! -> use arrays directly
  * unnecessary casts -> remove

Result: All 44 clippy warnings resolved. CI should now pass with -D warnings.

+++ nanopdf-rs/src/ffi/buffer.rs
@@ -845,7 +845,7 @@ mod tests {
     fn test_buffer_append_multiple() {
         let handle = fz_new_buffer(0, 0);
         for i in 0..100 {
-            fz_append_byte(0, handle, i as i32);
+            fz_append_byte(0, handle, i);
         }
         assert_eq!(fz_buffer_len(0, handle), 100);
         fz_drop_buffer(0, handle);
